SpD, multiprojection heads (#306)
### Objective
This PR seeks to implement post-attention hidden size projections used
to speculate tokens ahead of the base model. This PR contains three
primary components:

1. extending base model with multi-projection in `modeling_auto.py`
2. implementing multi-projections forward pass.
3. app demo of multi-projection model.

### Initial Implementation
Initial implementation gives the user the flexibility to define its own
projection architecture and pass it to `QEffAutoModelForCausalLM`.

Then, QEfficient simply attaches these projections to the model to be
used during the forward pass. The "attaching" of these projections is
done by using the `accelerate` library. I used this library because it
has a robust implementation to attach weights to an already existing
model. We can implement our own abstraction if needed, but first we must
agree on what the external API will be to the user.

> NOTE: Please keep in mind that to integrate medusa, similar changes
will be needed (instead of doing multiple hidden size projections,
medusa uses multiple `lm_heads` to speculate ahead of the base model).

---------

Signed-off-by: eplatero <quic_eplatero@quicinc.com>