Hash bug fix pr (#384)
Adding model_card_name to hash

Addition of model_card_name to hash helps to deal with error which
occurs when two models sharing the same architecture but have different
model card names and weights.

In future if any more parameters are required to make the hash unique,
they can be a part of the kwargs as implemeted here.

---------

Signed-off-by: Dipankar Sarkar <quic_dipankar@quicinc.com>