Fix for Replicate_kv_heads script (#412)
As we have moved to the transformer version 4.50.0, we have changed the
attention implementation and hence number of attention heads and hidden
size are no more attributes of this class.

We have added these parameters are optional arguments for the
replicate_kv_heads script. User can now pass them as arguments, if not
passed explicitly then it shall be picked from config.json file of the
model.

---------

Signed-off-by: Hem Agnihotri <hemagnih@qti.qualcomm.com>