Cleaned and Rebased PR for (#481) to change the hash creation module … (#537)
…for all models in Qefficient.

This PR contains changes made to the modelling_qeff, modeling_auto to
allow usage of certain export parameters and kwargs passed during model
creation. The hashing module is now made independant of the calling
class and the test scripts are updated accordingly to test for this
functionality. Added functionality to have an overarching parent
directory in cache to contain all different exported model configs
belonging to the same architecture. In case the architecture isn't
present in the config of the model, we instead proceed with
self.model_name based parent directory creation. Hash is now created
during export, so as to incorporate all the additional params needed for
unique hash creation thus, the test scripts have been modified to test
hashing functionalities accordingly. We maintain an Exclusion list of
params for kwargs to be discarded during hashing parameter selection.
We'll need to look into the alternate approach of maintaining an
Inclusion list instead. There was a comment to use MetaClasses to handle
raising a warning whenever someone loads a model without using
'from_pretrained' method but the current class architecture of VLMs and
SpeechSeq2Seq models don't allow for this, this use case will be handled
in a different PR.

---------

Signed-off-by: Dhiraj Kumar Sah <dhirajku@qti.qualcomm.com>