Unit Tests for On Device Sampling (#463)
This PR adds the following Unit Tests for On Device Sampling:
1. `test_sampler_transform`: Test if `SamplerTransform` adds nodes at
the output of a `QEffForCausalLM model` to enable the sampling of next
tokens at the device (instead of the host) and returns the next tokens
and/or probability distributions.
2. `test_greedy_sampling`: Test greedy sampling with QPC compiled with
and without On Device Sampling.
3. `test_random_sampling`: Test random sampling with QPC compiled with
and without On Device Sampling.

---------

Signed-off-by: quic-sanising <quic_sanising@quicinc.com>
Signed-off-by: sanising <sanising@qti.qualcomm.com>
Signed-off-by: Dhiraj Kumar Sah <dhirajku@qti.qualcomm.com>
Signed-off-by: Rishin Raj <rishinr@qti.qualcomm.com>
Co-authored-by: sanising <sanising@qti.qualcomm.com>
Co-authored-by: Dhiraj Kumar Sah <dhirajku@qti.qualcomm.com>
Co-authored-by: Rishin Raj <rishinr@qti.qualcomm.com>