Updated KV_pytorch and ORT inference for VLMs to incorporate image_idx (#557)
Updated the run_vlm_kv_model_on_pytorch and run_vlm_kv_model_on_ort
methods to run for the latest dual QPC setup. Along with the required
changes to be made in the Input Handler of VLMs.

Also updated the way head_dim is calculated for past_key_value creation
as certain models now provide specific head_dim. We fallback to previous
method if the parameter isn't found in the config.

Signed-off-by: Dhiraj Kumar Sah <dhirajku@qti.qualcomm.com>
Signed-off-by: abhishek-singh591 <sabhis@qti.qualcomm.com>