Added Decoder layer class in Qeff for granite (#628)
This PR introduces support for the Decoder layer class in Qeff for
granite, which is required for the subfunction. While there are
alternative approaches to achieve this, implementing it now ensures
future compatibility, particularly if we decide to add CB support for
Granite. In that case, the Qeff Granite Decoder layer will be necessary.

Signed-off-by: abhishek-singh591 <sabhis@qti.qualcomm.com>
Signed-off-by: Dhiraj Kumar Sah <dhirajku@qti.qualcomm.com>