Add automatic CCL list generation for prefill and decode when user does not provide lists (#663)
For using CCL feature during prefilling and decoding, the user needs to
pass two lists of context lengths to be used during these processes.
Here, we add an option to generate these lists automatically when they
are not provided by the user manually. These list generation is suitable
for a general-purpose application which considers both prefilling and
decoding processes and generates CCL lists for both.

---------
Signed-off-by: Rishin Raj <rishinr@qti.qualcomm.com>
Signed-off-by: Vahid Janfaza <vjanfaza@qti.qualcomm.com>
Signed-off-by: abhishek-singh591 <sabhis@qti.qualcomm.com>